+++ example/src/Settings/GiphyDialogSettings.tsx
