+++ example/src/Settings/InputCard.tsx
