+++ example/src/Settings/SwitchCard.tsx
