+++ example/src/Settings/index.ts
